+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/AppStoreRefactored/FormSchema/types.ts
@@ -1,0 +1,5 @@
+import * as yup from "yup";
+import { localisationFormSchema, mainAppStoreFormSchema } from "./form-schema";
+
+export type LocalisationFormSchema = yup.Asserts<typeof localisationFormSchema>;
+export type AppStoreFormValues = yup.Asserts<typeof mainAppStoreFormSchema>;
